+++ router/src/main/java/com/chenenyu/router/IRouter.java
@@ -1,5 +1,6 @@
 package com.chenenyu.router;
 
+import android.app.ActivityOptions;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
@@ -16,8 +17,14 @@ import android.support.v4.app.Fragment;
  * Created by chenenyu on 2017/3/31.
  */
 public interface IRouter {
+    /**
+     * Entrance.
+     */
     IRouter build(Uri uri);
 
+    /**
+     * Route request callback.
+     */
     IRouter callback(RouteCallback callback);
 
     /**
@@ -72,9 +79,9 @@ public interface IRouter {
     IRouter anim(@AnimRes int enterAnim, @AnimRes int exitAnim);
 
     /**
-     * {@link ActivityOptionsCompat}.
+     * {@link ActivityOptions#toBundle()} and {@link ActivityOptionsCompat#toBundle()}.
      */
-    IRouter activityOptions(ActivityOptionsCompat activityOptions);
+    IRouter activityOptionsBundle(Bundle activityOptionsBundle);
 
     /**
      * Skip all the interceptors.
@@ -91,9 +98,21 @@ public interface IRouter {
      */
     IRouter addInterceptors(String... interceptors);
 
-    Intent getIntent(Context context);
+    /**
+     * Get an intent instance.
+     *
+     * @param source Activity or Fragment instance.
+     * @return {@link Intent} instance.
+     */
+    Intent getIntent(Object source);
 
-    Object getFragment(Context context);
+    /**
+     * Get a fragment instance.
+     *
+     * @param source Activity or Fragment instance.
+     * @return {@link Fragment} or {@link android.app.Fragment} instance.
+     */
+    Object getFragment(Object source);
 
     void go(Context context, RouteCallback callback);
 
@@ -106,26 +125,4 @@ public interface IRouter {
     void go(android.app.Fragment fragment, RouteCallback callback);
 
     void go(android.app.Fragment fragment);
-
-    /**
-     * Call method by a {@link MethodCallable} instance. The method can be static or not.
-     * If the method has parameters, each param's type must be {@link String} and
-     * all the params must be annotated by {@link com.chenenyu.router.annotation.InjectParam}.
-     *
-     * @param context  Context
-     * @param callable A {@link MethodCallable} instance.
-     * @return True if find the method and invoke successfully, false otherwise.
-     */
-    boolean go(Context context, MethodCallable callable);
-
-    /**
-     * Call method by a subclass of {@link MethodCallable}. <strong>The method must be static.</strong>
-     * If the method has parameters, each param's type must be {@link String} and
-     * all the params must be annotated by {@link com.chenenyu.router.annotation.InjectParam}.
-     *
-     * @param context Context
-     * @param clz     Class that implements {@link MethodCallable} interface.
-     * @return True if find the method and invoke successfully, false otherwise.
-     */
-    boolean go(Context context, Class<? extends MethodCallable> clz);
 }